requirements.txt
mysql-connector-python
DBUtils
PyJWT
oss2
python-multipart
fastapi
uvicorn
pydantic-settings